Move HTTPClient to own file
